requirements.txt
streamlit
pandas
matplotlib
python-dotenv
openai
requests
streamlit-authenticator
cachetools
apscheduler
